Have Features Webclimbing, Catapults, SpawnPoints
The player is able to climb while in webs but the movement logic should be edited it still doesn't feel smooth.
The catapult webs are implemented and work well and are set up so that the code can be easily tweaked to have the catapult behave in a different manner(shoot the player straight upwards, shoot them in a higher arc, shoot them in a line) easily.
Spawn points work but need to be tweaked as well. I am not disabeling the spawn triggers so if you fall and hit an older trigger you can set that as the active spawn point. For level testing that's not great but for the final game depending on how levels are designed that might be nice. I went into the input manager and added a Respawn input tied to the R key but I didn't mess with any controller input so I'm pretty sure it only works with keyboard and we might have problems if we are using a controller.

Bug: If you are in a web you can not respawn because of how I am calling the respawn method but that should be an easy fix.