Have Features Webclimbing, Catapults, SpawnPoints
The player is able to climb while in webs but the movement logic should be edited it still doesn't feel smooth.
The catapult webs are implemented and work well and are set up so that the code can be easily tweaked to have the catapult behave in a different manner(shoot the player straight upwards, shoot them in a higher arc, shoot them in a line) easily.
Spawn points work but need to be tweaked as well. I am not disabeling the spawn triggers so if you fall and hit an older trigger you can set that as the active spawn point. For level testing that's not great but for the final game depending on how levels are designed that might be nice. I went into the input manager and added a Respawn input tied to the R key but I didn't mess with any controller input so I'm pretty sure it only works with keyboard and we might have problems if we are using a controller.

Bug: If you are in a web you can not respawn because of how I am calling the respawn method but that should be an easy fix.
Fixed some stuff

Fixe more stuff
Fix more stuff

fff
All Changes Made

So I changed a couple of variables in the player controller to public but those are all marked. In the input manager in unity I created a Grab(w), Drop(s), and Action(left shift) axis. So now when going to grab a web you must click w to grab it and can drop from it either by jumping or by holding s and leftshift. The player also has a respawn axis which is R. In the player controller if they hit R their current transform is set to a spawnTransform which is equal to the spawnpoints position.

All of the code for web movement is in the StickyWebController class and movement is fixed up and feeling much better.

I removed the slingshots because I was having trouble getting them to do exactly what I wanted them to do.

The spawn point has been changed all of the logic controlling how it moves around is cleaned up. Now when the player hits a SpawnTrigger which is an invisible volume, the spawnpoint moves to the triggers location. The trigger has a switch which moves it to a couple of preset locations, so now one trigger controls where the spawn point moves instead of a bunch of seperate triggers.
AI Scene Created


Bug Hive AI working

I need to tune it so the bugs spawn and move a bit because right now they all are overlapping.
Otherwise when the player gets within range of the leaderbug it runs back to the hive to get help from it's followers which then buzz around it. 
When it has it's followers it chases the player