Got rid of Most 2D colliders In EveryScene. Otherwisee there will be a ton of merge conflicts (#41)
So as the discription says. I purged most of the 2d colliders and rigidbodies. 
I also organized the scripts into seperate folders, there is an Enemies, Environment, NPC's, Pathing, and Player folder for all of our scripts to be organized in.

NOTE: There are still a couple of 2d rigidbodies and colliders in some scenes because some people are using scripts that depend on rigid bodies and colliders so rather than remove your scripts and cause any possible confusion I left those in. So that is the jumping enemy and the falling platforms. The hugger AI uses a deprecated version of Pawn AABB but that is still on there. 

I added PawnABB3d To the player prefab and all the players in your scenes so we shouldn't get the million null reference exception errors.

Otherwise with the exception of what was stated above all environment objects have the colliders and rigidbodies removed. This is so when we eventually go to merge if we don't delete all the other scenes then we will have a ton of merge conflicts. This way if this is merged back into master everyone can start out clean.

Oh I also deleted an empty scene and my test AI scene.
Forgot A collider and added something to the player prefab