Big changes to curves! #27
Curved sections of the path are now CIRCLES instead of cubic BEZIERS. The math is a bit more complicated, but the results are great!

The PathNode class saw significant refactoring, especially Constrain() and CacheData(). I also added a few utilities like CurveAngleTo(), ProjectOnCurve(), and AngleWrapFromTo().

This takes care of issue #27.