Deleted old Hive Controller, Changed fly behavior
Edited fly behavior to be a bit more stable, deleted the old hive controller so there is no confusion as to which one to attach.

Added another child collider to flies

So this second collider doesn't effect collision with the fly it's only job is to switch the player state. This still has to be implemented but the plan is once the player is hit by this trigger their state changes so they  slow down and can't jump as high
Fly Behavior adjusted

Flies now have a bounding box that slows the player to represent them being bugged(haha get it?) 
Player StateStateBugged implemented, Flies refined further

So now when the flies are harrasing the player a counter starts and the player is shunted into a state where they move slower and can't jump except for little hops everynow and then, they then have to lure the flies to a web to trap them, afterwards they are able to jump and move normally.