NPC Vision Gizmos
Added a script that draws a wireframe around an NPC representing it's "vision" or attack range.

Drop script onto any enemy, on the same sub-object that has whatever collision volume is detecting the player.

Script adds a dropdown that allows you to chose to view the collision range for this object only, all objects, or none at all.

Helps to see enemies range when level designing.

Can also be turned off and on in the gizmo tray of teh scene window.