Moving Platforms working!
Removed old platform player moving logic. 
PlatformController now only moves the platform. The StickyVolume moves the player.
Added a MovingPlatform prefab to the prefabs folder.
ADDITIONAL: If the platform is moving at high rates of speed, the player has the potential to fall through it
Platform movement isn't constrained to the paths, but the player still is. If the platform leaves the path, the player will fall off of it